+++ server/chalicelib/s3_historical.py
@@ -3,7 +3,7 @@ from chalicelib import s3
 from chalicelib.constants import EVENT_ARRIVAL, EVENT_DEPARTURE
 
 import itertools
-import dates
+import chalicelib.date_utils as date_utils
 
 
 def pairwise(iterable):
@@ -44,15 +44,15 @@ def dwells(stop_ids, sdate, edate):
             and maybe_a_departure["event_type"] in EVENT_DEPARTURE
             and maybe_an_arrival["trip_id"] == maybe_a_departure["trip_id"]
         ):
-            dep_dt = dates.parse_event_date(maybe_a_departure["event_time"])
-            arr_dt = dates.parse_event_date(maybe_an_arrival["event_time"])
+            dep_dt = date_utils.parse_event_date(maybe_a_departure["event_time"])
+            arr_dt = date_utils.parse_event_date(maybe_an_arrival["event_time"])
             delta = dep_dt - arr_dt
             dwells.append(
                 {
                     "route_id": maybe_a_departure["route_id"],
                     "direction": int(maybe_a_departure["direction_id"]),
-                    "arr_dt": dates.return_formatted_date(arr_dt),
-                    "dep_dt": dates.return_formatted_date(dep_dt),
+                    "arr_dt": date_utils.return_formatted_date(arr_dt),
+                    "dep_dt": date_utils.return_formatted_date(dep_dt),
                     "dwell_time_sec": delta.total_seconds(),
                 }
             )
@@ -72,8 +72,8 @@ def headways(stop_ids, sdate, edate):
             # Here, we should skip the headway
             # (though if trip_id is empty, we don't know).
             continue
-        this_dt = dates.parse_event_date(this["event_time"])
-        prev_dt = dates.parse_event_date(prev["event_time"])
+        this_dt = date_utils.parse_event_date(this["event_time"])
+        prev_dt = date_utils.parse_event_date(prev["event_time"])
         delta = this_dt - prev_dt
         headway_time_sec = delta.total_seconds()
 
@@ -119,8 +119,8 @@ def travel_times(stops_a, stops_b, sdate, edate):
         dep = departure["event_time"]
         arr = arrival["event_time"]
 
-        dep_dt = dates.parse_event_date(dep)
-        arr_dt = dates.parse_event_date(arr)
+        dep_dt = date_utils.parse_event_date(dep)
+        arr_dt = date_utils.parse_event_date(arr)
         delta = arr_dt - dep_dt
         travel_time_sec = delta.total_seconds()
 
@@ -141,8 +141,8 @@ def travel_times(stops_a, stops_b, sdate, edate):
             {
                 "route_id": departure["route_id"],
                 "direction": int(departure["direction_id"]),
-                "dep_dt": dates.return_formatted_date(dep_dt),
-                "arr_dt": dates.return_formatted_date(arr_dt),
+                "dep_dt": date_utils.return_formatted_date(dep_dt),
+                "arr_dt": date_utils.return_formatted_date(arr_dt),
                 "travel_time_sec": travel_time_sec,
                 "benchmark_travel_time_sec": benchmark,
             }
